timer: add comments for alarm module

+++ adaptation/platform/system.h
@@ -12,16 +12,29 @@
 extern "C" {
 #endif
 
+/**
+ * This method initialize the platform.
+ */
 void ndn_platform_init(void);
 
+/**
+ * This method de-initialize the platform.
+ */
 void ndn_platform_deinit(void);
 
+/**
+ * This method delay the platform in millisecond, like no op.
+ * @param delay. Time to delay in millisecond.
+ */
 void ndn_platform_delay_ms(uint32_t delay);
 
-uint32_t ndn_platform_current_time(void);
+/**
+ * This static method returns the current time in milliseconds.
+ * @return The current time in milliseconds.
+ */
+uint64_t ndn_platform_current_time(void);
 
 
 #ifdef __cplusplus
 } // end of extern "C"
 #endif
-
